+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-### 2018-8-14
+### 2018-08-27
+- (2.2.4)`getApp`新增`allowDefault`参数
+- (2.2.3)自定义组件新增了`lifetimes`、`pageLifetimes`、`definitionFilter`字段
+- (2.2.3)`backgroundAudioManager`新增`onSeeking`和`onSeeked`方法
+- (2.2.3)新增`wx.nextTick`方法
+### 2018-08-14
 - 更正`SelectorQuery`的`in`方法参数类型和返回类型